+++ tests/test_main.py
@@ -7,7 +7,7 @@ from kloch.__main__ import main
 
 
 def test__main__logging(monkeypatch, tmp_path: Path):
-    log_path = tmp_path / "kloch.log"
+    log_path = tmp_path / "logs" / "kloch.log"
     monkeypatch.setenv(kloch.Environ.CONFIG_CLI_LOGGING_PATHS, str(log_path))
     assert not log_path.exists()
     with pytest.raises(SystemExit):