+++ tests/test_profile.py
@@ -6,12 +6,12 @@ def test__EnvironmentProfile__merging():
     profile1 = EnvironmentProfile(
         identifier="knots",
         version="0.1.0",
-        base=None,
+        inherit=None,
         launchers=LauncherSerializedDict(
             {
                 "rezenv": {
-                    "+=config": {"exclude": "whatever"},
-                    "requires": {
+                    "config": {"exclude": "whatever"},
+                    "==requires": {
                         "echoes": "2",
                         "maya": "2023",
                     },
@@ -27,20 +27,20 @@ def test__EnvironmentProfile__merging():
     profile2 = EnvironmentProfile(
         identifier="knots:echoes",
         version="0.1.0",
-        base=profile1,
+        inherit=profile1,
         launchers=LauncherSerializedDict(
             {
                 "+=rezenv": {
-                    "config": {"include": "yes"},
-                    "+=requires": {
+                    "==config": {"include": "yes"},
+                    "requires": {
                         "-=maya": "_",
                         "-=notAdded": "_",
                         "added": "1.2",
                     },
                     "+=tests": {
-                        "+=foo": [4, 5, 6],
+                        "foo": [4, 5, 6],
                         "+=new-echoes-key": {"working": True},
-                        "deeper!": {"+=as deep": [0, 0]},
+                        "==deeper!": {"+=as deep": [0, 0]},
                     },
                 }
             }
@@ -49,15 +49,15 @@ def test__EnvironmentProfile__merging():
     result = profile2.get_merged_profile().launchers
     expected = {
         "+=rezenv": {
-            "config": {"include": "yes"},
-            "+=requires": {
+            "==config": {"include": "yes"},
+            "requires": {
                 "echoes": "2",
                 "added": "1.2",
             },
             "+=tests": {
-                "+=foo": [1, 2, 3, 4, 5, 6],
+                "foo": [1, 2, 3, 4, 5, 6],
                 "+=new-echoes-key": {"working": True},
-                "deeper!": {"+=as deep": [0, 0]},
+                "==deeper!": {"+=as deep": [0, 0]},
             },
         },
         "testenv": {"command": "echo $cwd"},