+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2024-08-04
+
+### added
+
+- profile: new `!=` "if not exists" token
+- the log file parent directory is created if it doesn't exist
+- it's possible to add environment variable in the keys of the config file that are paths
+
+### changed
+
+- ! cli: `--profile_paths` renamed to `--profile_roots`
+- ! profile: the default merge rule is now "append" instead of "override"
+- ! profile: renamed the root `base` key to `inherit`
+- logs go up to 262144 bytes before being rotated
+
 ## [0.9.0] - 2024-07-03
 
 ### added
@@ -14,7 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### fixed
 
 - keys in config file have their type properly casted to the expected type
-
 
 
 ## [0.8.2] - 2024-06-29